move to svelte-kit
