+++ .gitignore
@@ -1,6 +1,15 @@
-/node_modules/
-/public/build/
-package-lock.json
-
 .DS_Store
-*.todo
+node_modules
+/build
+/dist
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
+
+/_dist
+/_package.json
+/.npmrc
