--- index.js
@@ -1,5 +1,0 @@
-import Svelecte from './src/Svelecte.svelte';
-
-export default Svelecte;
-export { addFormatter, config, TAB_SELECT_NAVIGATE } from './src/Svelecte.svelte';
-export { registerSvelecte } from './src/lib/custom-element';
