--- item.js
@@ -1,5 +1,0 @@
-import CloseButton from './src/components/ItemClose.svelte';
-import { highlightSearch } from './src/lib/utils';
-import itemActions from './src/actions';
-
-export { CloseButton, itemActions, highlightSearch }
