+++ src/lib/index.js
@@ -1,0 +1,5 @@
+import Svelecte from './Svelecte.svelte';
+
+export default Svelecte;
+export { addFormatter, config, TAB_SELECT_NAVIGATE } from './Svelecte.svelte';
+export { registerSvelecte } from './lib/custom-element.js';
