+++ src/lib/item.js
@@ -1,0 +1,5 @@
+import CloseButton from './src/components/ItemClose.svelte';
+import { highlightSearch } from './src/lib/utils.js';
+import itemActions from './src/actions.js';
+
+export { CloseButton, itemActions, highlightSearch }
